entire logic and API updated due to critical logical bug found
ver 2.0 was tested when requiring module and required module are at the
same folder. however, if they are not, since the require function that
is used is from a different scope, it not would find the desired module to
be required (because every require function is wired with a different
path). that critical issue required total logical and API update.